Refactored Charactrer View
