+++ SolidCharacters/client/src/components/characters/view/sheet/sheet.module.scss
@@ -488,10 +488,10 @@ $mobile: 900px;
 }
 
 .deathRow {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-1);
-  margin-bottom: 6px;
+  // display: flex;
+  // align-items: center;
+  // gap: var(--spacing-1);
+  // margin-bottom: 6px;
 }
 
 .deathRowLabel {
